+++ src/main/java/org/scion/jpan/Scmp.java
@@ -475,117 +475,21 @@ public class Scmp {
   }
 
   /**
-   * Create a channel for sending SCMP requests.
+   * Create a channel for sending (blocking) SCMP requests.
    *
    * @return New SCMP channel
-   * @deprecated Experimental, subject to change.
    */
-  @Deprecated
-  public static ScmpChannel2 createAsyncChannel() throws IOException {
-    return new ScmpChannel2();
+  public static ScmpBlockingSender.Builder newBlockingSenderBuilder() {
+    return ScmpBlockingSender.newBuilder();
   }
 
   /**
-   * Create a channel for sending SCMP requests.
-   *
-   * @param listeningPort Local port to listen for SCMP requests.
-   * @return New SCMP channel
-   * @deprecated Experimental, subject to change.
-   */
-  @Deprecated
-  public static ScmpChannel2 createAsyncChannel(int listeningPort) throws IOException {
-    return new ScmpChannel2(Scion.defaultService(), listeningPort);
-  }
-
-  /**
-   * Create a channel for sending SCMP requests.
-   *
-   * @param service the ScionService instance
-   * @param listeningPort Local port to listen for SCMP requests.
-   * @return New SCMP channel
-   * @deprecated Experimental, subject to change.
-   */
-  @Deprecated
-  public static ScmpChannel2 createAsyncChannel(ScionService service, int listeningPort)
-      throws IOException {
-    return new ScmpChannel2(service, listeningPort);
-  }
-
-  /**
-   * Create a channel for sending SCMP requests.
+   * Create a channel for sending non-blocking SCMP requests.
    *
    * @return New SCMP channel
-   * @deprecated Experimental, subject to change.
    */
-  @Deprecated
-  public static ScmpChannel3 createAsyncChannel3() throws IOException {
-    return new ScmpChannel3();
-  }
-
-  /**
-   * Create a channel for sending SCMP requests.
-   *
-   * @param listeningPort Local port to listen for SCMP requests.
-   * @return New SCMP channel
-   * @deprecated Experimental, subject to change.
-   */
-  @Deprecated
-  public static ScmpChannel3 createAsyncChannel3(int listeningPort) throws IOException {
-    return new ScmpChannel3(Scion.defaultService(), listeningPort);
-  }
-
-  /**
-   * Create a channel for sending SCMP requests.
-   *
-   * @param service the ScionService instance
-   * @param listeningPort Local port to listen for SCMP requests.
-   * @return New SCMP channel
-   * @deprecated Experimental, subject to change.
-   */
-  @Deprecated
-  public static ScmpChannel3 createAsyncChannel3(ScionService service, int listeningPort)
-      throws IOException {
-    return new ScmpChannel3(service, listeningPort);
-  }
-
-  /**
-   * Create a channel for sending SCMP requests.
-   *
-   * @return New SCMP channel
-   * @deprecated Experimental, subject to change.
-   */
-  @Deprecated
-  public static ScmpSender createSender(ScmpSender.ScmpResponseHandler handler) throws IOException {
-    // TODO fix this: 12345 is bad.
-    return new ScmpSender(Scion.defaultService(), 12345, handler);
-  }
-
-  /**
-   * Create a channel for sending SCMP requests.
-   *
-   * @param listeningPort Local port to listen for SCMP requests.
-   * @return New SCMP channel
-   * @deprecated Experimental, subject to change.
-   */
-  @Deprecated
-  public static ScmpSender createSender(ScmpSender.ScmpResponseHandler handler, int listeningPort)
-      throws IOException {
-    return new ScmpSender(Scion.defaultService(), listeningPort, handler);
-  }
-
-  /**
-   * Create a channel for sending SCMP requests.
-   *
-   * @param service the ScionService instance
-   * @param listeningPort Local port to listen for SCMP requests.
-   * @return New SCMP channel
-   * @deprecated Experimental, subject to change.
-   */
-  @Deprecated
-  public static ScmpSender createSender(
-      ScmpSender.ScmpResponseHandler handler, ScionService service, int listeningPort)
-      throws IOException {
-    return new ScmpSender(service, listeningPort, handler);
+  public static ScmpSender.Builder newSenderBuilder(ScmpSender.ScmpResponseHandler handler) {
+    return ScmpSender.newBuilder(handler);
   }
 
   /**
@@ -593,9 +497,8 @@ public class Scmp {
    * back.
    *
    * @return New SCMP responder
-   * @deprecated Experimental, subject to change.
    */
-  public static ScmpResponder createResponder() throws IOException {
-    return new ScmpResponder(Scion.defaultService(), Constants.SCMP_PORT);
+  public static ScmpResponder.Builder newResponderBuilder() {
+    return ScmpResponder.newBuilder();
   }
 }